Rename and daemonize the bot
Henceforth, this IRC Bot is called Snap. Rather than live off of a set
configuration for myself, it is now available as a clonable and
configurable bot for others who wish simply to have a bot they can start
up and not have to develop it themselves. This is a bootstrapped bot
using Cinch and will be updated with plugins as I develop them.

Signed-off-by: David Celis <david@davidcelis.com>